Separate functions into unique file
